refactor: move `OccurrenceAnalyzer` to `optimizer` package (#10384)
