+++ main/src/ca/uwaterloo/flix/language/phase/optimizer/Optimizer.scala
@@ -3,7 +3,6 @@ package ca.uwaterloo.flix.language.phase.optimizer
 import ca.uwaterloo.flix.api.Flix
 import ca.uwaterloo.flix.language.ast.{MonoAst, OccurrenceAst}
 import ca.uwaterloo.flix.language.dbg.AstPrinter.DebugMonoAst
-import ca.uwaterloo.flix.language.phase.OccurrenceAnalyzer
 import ca.uwaterloo.flix.util.ParOps
 
 object Optimizer {
